+++ .goreleaser.yml
@@ -12,6 +12,13 @@ builds:
       - linux
     goarch:
       - amd64
+    ldflags:
+      - -s -w
+      - -X main.version={{.Version}}
+      - -X main.commit={{.Commit}}
+      - -X main.date={{.Date}}
+      - -X main.builtBy=goreleaser
+      - -extldflags "-static"
   - id: tarian-cluster-agent
     main: ./cmd/tarian-cluster-agent/
     binary: tarian-cluster-agent
@@ -21,6 +28,13 @@ builds:
       - linux
     goarch:
       - amd64
+    ldflags:
+      - -s -w
+      - -X main.version={{.Version}}
+      - -X main.commit={{.Commit}}
+      - -X main.date={{.Date}}
+      - -X main.builtBy=goreleaser
+      - -extldflags "-static"
   - id: tarian-node-agent
     main: ./cmd/tarian-node-agent/
     binary: tarian-node-agent
@@ -33,8 +47,7 @@ builds:
     flags:
       - -tags=netgo
     ldflags:
-      - -s
-      - -w
+      - -s -w
       - -X main.version={{.Version}}
       - -X main.commit={{.Commit}}
       - -X main.date={{.Date}}
@@ -49,6 +62,13 @@ builds:
       - linux
     goarch:
       - amd64
+    ldflags:
+      - -s -w
+      - -X main.version={{.Version}}
+      - -X main.commit={{.Commit}}
+      - -X main.date={{.Date}}
+      - -X main.builtBy=goreleaser
+      - -extldflags "-static"
   - id: tarianctl
     main: ./cmd/tarianctl/
     binary: tarianctl
@@ -58,6 +78,13 @@ builds:
       - linux
     goarch:
       - amd64
+    ldflags:
+      - -s -w
+      - -X main.version={{.Version}}
+      - -X main.commit={{.Commit}}
+      - -X main.date={{.Date}}
+      - -X main.builtBy=goreleaser
+      - -extldflags "-static"
 
 dockers:
   - ids:
@@ -120,18 +147,18 @@ dockers:
     - "--label=org.opencontainers.image.revision={{.FullCommit}}"
     - "--label=org.opencontainers.image.version={{.Version}}"
     - "--label=org.opencontainers.image.source={{ .GitURL }}"
-  - ids:
-    - tarianctl
-    goos: linux
-    goarch: amd64
-    image_templates:
-    - '{{ .Env.CONTAINER_REGISTRY }}/tarianctl:{{ if .IsSnapshot }}latest{{ else }}{{ trimprefix .Tag "v" }}{{ end }}'
-    dockerfile: Dockerfile-tarianctl
-    skip_push: false
-    build_flag_templates:
-    - "--pull"
-    - "--label=org.opencontainers.image.created={{.Date}}"
-    - "--label=org.opencontainers.image.title={{.ProjectName}}"
-    - "--label=org.opencontainers.image.revision={{.FullCommit}}"
-    - "--label=org.opencontainers.image.version={{.Version}}"
-    - "--label=org.opencontainers.image.source={{ .GitURL }}"
+  # - ids:
+  #   - tarianctl
+  #   goos: linux
+  #   goarch: amd64
+  #   image_templates:
+  #   - '{{ .Env.CONTAINER_REGISTRY }}/tarianctl:{{ if .IsSnapshot }}latest{{ else }}{{ trimprefix .Tag "v" }}{{ end }}'
+  #   dockerfile: Dockerfile-tarianctl
+  #   skip_push: false
+  #   build_flag_templates:
+  #   - "--pull"
+  #   - "--label=org.opencontainers.image.created={{.Date}}"
+  #   - "--label=org.opencontainers.image.title={{.ProjectName}}"
+  #   - "--label=org.opencontainers.image.revision={{.FullCommit}}"
+  #   - "--label=org.opencontainers.image.version={{.Version}}"
+  #   - "--label=org.opencontainers.image.source={{ .GitURL }}"
