+++ cmd/version.go
@@ -1,0 +1,11 @@
+package version
+
+var versionStr string
+
+func SetVersion(ver string) {
+	versionStr = ver
+}
+
+func GetVersion() string {
+	return versionStr
+}
